% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subreddit_info.R
\name{get_subreddit_info}
\alias{get_subreddit_info}
\title{Retrieve Subreddit Information}
\usage{
get_subreddit_info(subreddit, access_token)
}
\arguments{
\item{subreddit}{A string specifying the name of the subreddit.}

\item{access_token}{A string representing the access token obtained via Reddit API authentication.}
}
\value{
A data frame containing details of the subreddit, including:
\item{title}{The name of the subreddit.}
\item{description}{The public description of the subreddit.}
\item{subscribers}{The number of subscribers to the subreddit.}
\item{active_users}{The number of active users in the subreddit.}
\item{created_utc}{The creation date of the subreddit (UTC).}
\item{over18}{Boolean indicating whether the subreddit is NSFW.}
\item{community_type}{The type of subreddit (e.g., public, private, restricted).}
\item{language}{The language setting of the subreddit.}
\item{icon_img}{URL of the subreddit’s icon image.}
\item{banner_img}{URL of the subreddit’s banner image.}
\item{primary_color}{Primary color of the subreddit’s theme.}
\item{allow_images}{Boolean indicating whether image posts are allowed.}
\item{allow_videos}{Boolean indicating whether video posts are allowed.}
\item{is_default}{Boolean indicating whether this is a default subreddit.}
\item{quarantine}{Boolean indicating whether the subreddit is quarantined.}
\item{submission_type}{The type of submissions allowed (e.g., "any", "link", "self").}
\item{wiki_enabled}{Boolean indicating whether the subreddit’s wiki is enabled.}
}
\description{
This function fetches detailed information about a specified subreddit using the Reddit API.
}
\examples{
\dontrun{
access_token <- reddit_auth("your_client_id", "your_client_secret", "your_username", "your_password")
subreddit_info <- get_subreddit_info("programming", access_token, username = "your_username")
print(subreddit_info)
}

}
