% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_subreddit_posts.R
\name{get_subreddit_posts}
\alias{get_subreddit_posts}
\title{Retrieve Hot Posts from a Subreddit}
\usage{
get_subreddit_posts(subreddit, access_token, limit, username)
}
\arguments{
\item{subreddit}{A string specifying the subreddit name (without the "/r/").}

\item{access_token}{A string representing the access token obtained via Reddit API authentication.}

\item{limit}{An integer specifying the number of posts to retrieve.}

\item{username}{A string representing the Reddit username used for the User-Agent header.}
}
\value{
A data frame containing details of the retrieved posts, including:
\item{id}{The post ID.}
\item{Title}{The title of the post.}
\item{Author}{The Reddit username of the post author.}
\item{Subreddit}{The name of the subreddit.}
\item{Upvotes}{The number of upvotes the post received.}
\item{Comments}{The number of comments on the post.}
\item{URL}{The URL of the post.}
\item{Created}{The timestamp of when the post was created.}
}
\description{
This function fetches the hottest posts from a specified subreddit using the Reddit API.
}
\examples{
\dontrun{
access_token <- reddit_auth("your_client_id", "your_client_secret", "your_username", "your_password")
posts <- get_subreddit_posts("rstats", access_token, limit = 5, username = "your_username")
print(posts)
}

}
