% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reddit_auth.R
\name{reddit_auth}
\alias{reddit_auth}
\title{Authenticate with the Reddit API}
\usage{
reddit_auth(client_id, client_secret, username, password)
}
\arguments{
\item{client_id}{A string representing the Reddit API client ID.}

\item{client_secret}{A string representing the Reddit API client secret.}

\item{username}{A string representing the Reddit account username.}

\item{password}{A string representing the Reddit account password.}
}
\value{
A string containing the access token if authentication is successful.
}
\description{
This function authenticates a user with the Reddit API using OAuth2 password credentials
and retrieves an access token.
}
\examples{
\dontrun{
token <- reddit_auth("your_client_id", "your_client_secret", "your_username", "your_password")
}

}
